Provide automatic plugin instantiation order
This patch improves the plugin loader algorithm by automatically
resolving the instantiation order of plugins based on their
dependencies. The loader builds a dependency graph and follows the
topological sorting of the graph to find the correct order of plugin
instantiation.

Now, each plugin can defines the optional class attribute
`OBJECTS_PROVIDED`, to provide the list of injected objects in the
plugin loading context, this attribute is used to solve object
dependency between plugins. The new objects `PluginRequirement`,
`ObjectRequirement` and `ContextObject` must be used to write plugin
dependency definition. The new Python module `plugin.context_config`
defines an unified definition of the plugin loading context that can be
used to define plugin dependencies which refer to a global common
definition.

This improvement is based on the builtin Python graphlib introduced in
Python 3.9, so this patch upgrades the Python dependency from 3.8 to
3.9.

Previously, the plugin system instantiated each plugin according to a
priority number hardcoded into its implementation. This design was
chosen for rapid prototyping of the plugin loader. Manual maintenance of
each plugin's priority number according to its dependencies was very
difficult, error prone, and should be revised when a new plugin was
released to avoid an instantiation order error due to a missing
dependency in the loading context. This patch solves all these problems.