Rename the emu option of fiit command to run
The fiit command has been modified to rename the `emu` option to `run`.
Fiit stands for Firmware Instrumentation and Introspection Tools, so
firmware emulation (`emu`) can be seen as a specific case of a group of
technics for introspecting firmware internals.

The command `fiit emu` takes a configuration file as argument and simply
call the plugin loader to performs firmware instrumentation and/or
analysis. The name of this option is confusing when order technics than
emulation is used to perform firmware analysis.