Upgrade the cmsis-svd Python dependency to v0.6
This patch updates the 'cmsis-svd' Python dependency to version 0.6.
Now, access to peripheral and register data structures is done via the
new API methods `get_peripherals()` and `get_registers()`, which allow
all elements to be collected regardless of the specific distribution of
these structures in the cmsis-svd structure tree. The design of the
`SVDLoader` class has been modified to remove unnecessary wrapper code,
which was written to handle future ways of loading a cmsis-svd file, but
abandoned for the moment.

Previously, device and register data structures were retrieved by
iterating over the Python data structure representing the cmsis-svd xml
file. Due to the different structure of the cmsis-svd file permitted by
the cmsis-svd file format, this could result in a partial collect of
peripherals and registers. This upgrade to the version 0.6 does not
completely solve the problem, and devices and/or registers may still be
missing if the cmsis-svd file is too complex, but this fix a number of
cases. Also, version 0.6 fix a high number of issues and provide new
features, for more details see the following commit
cmsis-svd/cmsis-svd@c8f4eb8.

This upgrade follows the merge of the long pending pull request
cmsis-svd/cmsis-svd#185 proposed by @VincentDary, the creator of this
project firmware-crunch/fiit.