Decouple the shell thread and the emulation thread
This patch merges the code of the `EmulatorShell` and `CustomShell`
classes in the new `Shell` class. The code related to the emulation
thread management has been removed, this factoring decouple the shell
and the emulation background task. The `UnicornEmulator` class provides
a new method to start the emulation in a thread, and the `%emu_start`
command of the `UnicornEmulatorFrontend` uses this feature to start the
emulation. The `Shell` class provides the new `suspend()` and `resume()`
methods for locking and unlocking the shell prompt, which are
independent of any parallel thread. The plugin 'plugin_emulator_shell'
has been renamed to 'plugin_shell'.

Providing an interface to only stop and resume the prompt in a generic
way without background thread management, allows to decouple the shell
and the parallel (emulation) thread, resulting in a better reuse of both
components. This decoupling is now necessary to develop future frontend
functionalities in the next project iteration steps. It is planned to
write features involving multiple stop and restart of the emulation
session. Decoupling the emulation thread and the shell thread allows a
better instrumentation of the emulation session when a shell is
configured and avoid future synchronisation issues or side effects
related to the management of the emulation thread by the shell.

In addition, the semantic of the `EmulatorShell` class was confusing.
This class was handled switching between two threads, the shell prompt
and a parallel task not specifically related to emulation, and the two
methods `resume_emu_exec()` and `start_emulation_thread() provided an
interface for suspending or resuming the shell prompt. The term
'emulation' is therefore confusing, since the parallel task can be
related to any parallel job. Changing the name of this class and methods
to generic names make the code easier to read.