Fix %cont and %step command of UnicornDbgFrontend
This patch prevents the `%cont` and `%step` commands of the
`UnicornDbgFrontend` class to suspend the shell prompt and to resume the
execution of the emulated code if the commands are not executed in the
context of a debug event.

Previously, no check was done, which led to lock indefinitely the
IPython/Jupyter kernel if one of these commands was executed outside the
context of a debugging event.