Improve fiit_console implementation
This patch improves the implementation of the custom Jupyter kernel
shell client 'fiit_console'. It provides exclusive access to the iopub
channel to avoid consuming external message [1] when running a cell. It
also implements custom processing and rendering of all external iopub
messages, which is part of a code factoring effort to make the frontend
easier to maintain.

The prompt loop task and the external message processing task are two
Python asynchronous tasks which consume messages in the 'iopub channel'
shared resource. The prompt loop task consumes iopub message in the
synchronous `run_cell()` method of the `ZMQTerminalInteractiveShell`
class, which is used to execute code in the remote Jupyter kernel. But,
the external message processing task consumes asynchronously cached
message in the `_post_jupyter_message_render()` method. This design can
potentially lead to inconsistent state if a cell is run while processing
external iopub message. A shared lock must be used around methods that
consume iopub messages to avoid any state inconsistency.

The prompt loop is interrupted when rendering external iopub message
related to an `execute_input` which match a sub set of input. In the
previous implementation, these messages were renderer with a custom
method for `execute_input` message type, and with the default rendering
for other message types. This design was based on the subclass of
`ZMQTerminalInteractiveShell.include_output` called in
`ZMQTerminalInteractiveShell.handle_iopub` to cache iopub messages and
delay rendering depending on some state based conditions. This
implementation works but is confused. Many states are used when at least
half of the states could be used to do the same job. Also, the rendering
path follows non explicit states that make the implementation difficult
to understand. By rewriting the rendering logic than use `handle_iopub`,
the code can be much easier to maintain.

[1] external message: refer to any received message which has a
    different session identifier in the `header.session` field. On the
    iopub channel, messages to all kernel client are published, so this
    field is used to filter external message.