Add machine, devices and cpu abstraction layer
This commit introduces a new architecture which brings lots of changes
in the global design of this project, and move this project from a
simple aggregation of tools to a structured and layered framework.

The new `machine` subpackage provide the material to create a machine
with sub-devices via the `Machine` and `MachineDevice` classes. The
concept of `Machine` and `Device` is left flexible, `Machine` should be
seen as DIY table onto which `Device` are screwed, and potentially
connected to each other with cables and other exotic mechanisms.

Moreover, the `machine` package provides a cpu abstraction via
`DeviceCpu` a subclass of `MachineDevice`, which provides basic cpu
introspection capabilities for registers, memory, and a hooking
interface for code execution, memory access and cpu exception. To
provide this abstraction, `DeviceCpu` use an implementation of the class
`Cpu` a pure abstract cpu interface in charge to provide the glue
between the high layer of the fiit framework and the physical or
emulated cpu.

The new `dev` subpackage provides reusable devices over the `machine`
abstraction. The new `emunicorn` subpackage implements the access to the
emulated Unicorn cpu.

A new plugin and file configuration interface provide a way to model
complex hardware scenario and introspection capabilities. A
configuration file can be used to create a machine plugin with its
devices, and next each introspection plugin can refer to one or many
devices specified in the machine plugin.

Almost the entire code of the project has been modified to support this
new architecture. Among these changes, many are related to the new cpu
registers and memory interfaces used by the cpu abstraction of the
`machine` package. Moreover, the plugin and configuration file interface
has been refunded to provide a machine and device based configuration.

Previously this project was highly focused on independent tools based on
the Unicorn cpu emulator library without any abstraction with this
library. Consequently the tools were strongly coupled with Unicorn and
unusable with another cpu emulator library, or in another context like a
physical cpu. Moreover, the projects and the plugin layer interface were
designed for a focus on only one cpu which was problematic to model a
complex scenario with several cpu and other independent hardware entity.